# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ BACKEND/src/service/productService.js
@@ -135,17 +135,22 @@ class ProductService {
   }
     
   // 관리자 상품 삭제
-  async deleteProduct(req, res){
+  async deleteProduct(req, res) {
     const { productId } = req.params;
     const product = await productModel.findByProductId(productId);
     if (!product) {
       throw new Error('해당 상품을 찾을 수 없습니다.');
     }
     // 이미지 파일 삭제
-    const existingImagePaths = product.imagePaths;
-    for (const imagePath of existingImagePaths) {
-      await fs.unlink(imagePath);
+    if (product.imagePaths && product.imagePaths.length > 0) {
+      const existingImagePaths = product.imagePaths;
+      for (const imagePath of existingImagePaths) {
+        if (fs.existsSync(imagePath)) { // 이미지 파일이 존재하는 경우에만 삭제
+          await fs.unlink(imagePath);
+        }
+      }
     }
+    // 상품 삭제
     return productModel.deleteById(productId);
   }
     
@@ -168,6 +173,11 @@ class ProductService {
   // 사용자가 인기상품 카테고리를 선택시 purchaseNum 10이상인 모든 상품을 조회
   getPopularProducts(req, res){
     return productModel.findByPurchaseNum(10);
+  }
+
+  // 사용자가 전체상품 카테고리를 선택시 모든 상품을 조회
+  getAllProducts(req, res){
+    return productModel.findByAll(0);
   }
     
   // 사용자가 상품 상제정보 확인
